requirements.txt
SpeechRecognition==3.10.4
keyboard==0.13.5
python-dotenv==1.0.0
google-generativeai==0.7.2
pywin32==308
numpy==1.24.3
openpyxl==3.1.2
